+++ src/features/handy/utils/getStrokeSpeedAndDistance.ts
@@ -13,6 +13,7 @@ export const fastestBps = maxStrokeSpeed / maxStrokeLength;
 export default function getStrokeSpeedAndDistance(bps: number) {
   let speed = maxStrokeSpeed;
   let stroke = maxStrokeLength;
+  const strokeZone = { min: 0, max: 100 }; // %
 
   // Decrease stroke speed to decrease BPS
   if (bps < fastestBps) {
@@ -22,11 +23,15 @@ export default function getStrokeSpeedAndDistance(bps: number) {
   // Shorten stroke distance to increase BPS
   if (bps > fastestBps) {
     stroke = maxStrokeSpeed / (bps * fastestBps);
+    strokeZone.min = Math.trunc(
+      ((maxStrokeLength - stroke) / maxStrokeLength) * 100
+    );
   }
 
   // Handy doesn't support decimal places with firmware 3+.
   return {
     speed: Math.trunc(speed),
     stroke: Math.trunc(stroke),
+    strokeZone,
   };
 }